MythicMobsSkillHandler now allows to make the "auto-targeting" optional on a per-ability basis.
For example, the ATTACK trigger includes the target of the ability being... the entity you attacked.
However, RIGHT_CLICK trigger does not include a target, so if the player was looking at an enemy, the auto-targeting will set this enemy as the target of the skill.

Now one can set in the YML file to make this particular ability leave the Mythic ability target-less:
  auto-target: false

_____
Also adds to the MythicLib config a value to set the default behaviour of all mythic abilities:

abilities:
  mythic-skill-autotarget: true